requirements.txt
azure-ai-textanalytics
azure-core
azure-cognitiveservices-speech
azure-cognitiveservices-vision-computervision
azure-identity
azure-storage-blob
boto3
Flask
Flask-Cors
guardrails-ai
langchain
langchain-openai
neo4j
openai
opencensus-ext-azure
pandas
plotly
pyodbc
python-docx
PyMuPDF
py2neo
redis

# *************Old Requirements************
# aiohappyeyeballs==2.4.0
# aiohttp==3.10.5
# aiosignal==1.3.1
# annotated-types==0.7.0
# annoy==1.17.3
# anyio==4.4.0
# arrow==1.3.0
# asgiref==3.8.1
# async-timeout==4.0.3
# attrs==24.2.0
# azure-ai-textanalytics==5.3.0
# azure-cognitiveservices-speech==1.40.0
# azure-cognitiveservices-vision-computervision==0.9.0
# azure-common==1.1.28
# azure-core==1.30.2
# azure-storage-blob==12.22.0
# bidict==0.23.1
# blinker==1.8.2
# boto3==1.35.11
# botocore==1.35.11
# certifi==2024.8.30
# cffi==1.17.1
# charset-normalizer==3.3.2
# click==8.1.7
# colorama==0.4.6
# coloredlogs==15.0.1
# cryptography==43.0.1
# dataclasses-json==0.6.7
# Deprecated==1.2.14
# diff-match-patch==20230430
# distro==1.9.0
# exceptiongroup==1.2.2
# Faker==25.9.2
# fastapi==0.115.0
# fastembed==0.3.6
# filelock==3.16.1
# Flask==3.0.3
# Flask-Cors==5.0.0
# Flask-SocketIO==5.3.6
# flatbuffers==24.3.25
# fqdn==1.5.1
# frozenlist==1.4.1
# fsspec==2024.9.0
# googleapis-common-protos==1.65.0
# griffe==0.36.9
# grpcio==1.66.1
# guardrails-ai==0.5.10
# guardrails-api-client==0.3.12
# guardrails_hub_types==0.0.4
# h11==0.14.0
# httpcore==1.0.5
# httpx==0.27.2
# huggingface-hub==0.25.0
# humanfriendly==10.0
# idna==3.8
# importlib_metadata==8.4.0
# interchange==2021.0.4
# isodate==0.6.1
# isoduration==20.11.0
# itsdangerous==2.2.0
# Jinja2==3.1.4
# jiter==0.5.0
# jmespath==1.0.1
# joblib==1.4.2
# jsonpatch==1.33
# jsonpointer==3.0.0
# jsonref==1.1.0
# jsonschema==4.23.0
# jsonschema-specifications==2023.12.1
# langchain==0.2.16
# langchain-community==0.2.17
# langchain-core==0.2.41
# langchain-text-splitters==0.2.4
# langsmith==0.1.125
# lark==1.1.9
# litellm==1.47.0
# loguru==0.7.2
# lxml==4.9.4
# markdown-it-py==3.0.0
# MarkupSafe==2.1.5
# marshmallow==3.22.0
# mdurl==0.1.2
# mmh3==4.1.0
# monotonic==1.6
# mpmath==1.3.0
# msrest==0.7.1
# multidict==6.1.0
# mypy-extensions==1.0.0
# nemoguardrails==0.9.1.1
# neo4j==5.24.0
# nest-asyncio==1.6.0
# networkx==3.3
# nltk==3.8.1
# numpy==1.26.4
# oauthlib==3.2.2
# onnx==1.16.2
# onnxruntime==1.19.2
# openai==1.47.0
# opentelemetry-api==1.27.0
# opentelemetry-exporter-otlp-proto-common==1.27.0
# opentelemetry-exporter-otlp-proto-grpc==1.27.0
# opentelemetry-exporter-otlp-proto-http==1.27.0
# opentelemetry-proto==1.27.0
# opentelemetry-sdk==1.27.0
# opentelemetry-semantic-conventions==0.48b0
# orjson==3.10.7
# packaging==24.1
# pandas==2.2.3
# pansi==2020.7.3
# pdfminer.six==20231228
# pdfplumber==0.11.4
# pillow==10.4.0
# plotly==5.24.1
# prompt_toolkit==3.0.47
# protobuf==4.25.5
# py2neo==2021.2.4
# pycparser==2.22
# pydantic==2.8.2
# pydantic_core==2.20.1
# pydash==7.0.7
# Pygments==2.18.0
# PyJWT==2.9.0
# PyMuPDF==1.24.9
# PyMuPDFb==1.24.9
# PyPDF2==3.0.1
# pypdfium2==4.30.0
# PyStemmer==2.2.0.1
# python-dateutil==2.9.0.post0
# python-docx==1.1.2
# python-dotenv==1.0.1
# python-engineio==4.9.1
# python-socketio==5.11.3
# pytz==2024.2
# PyYAML==6.0.2
# redis==5.0.8
# referencing==0.35.1
# regex==2023.12.25
# requests==2.32.3
# requests-oauthlib==2.0.0
# rfc3339-validator==0.1.4
# rfc3987==1.3.8
# rich==13.8.1
# rpds-py==0.20.0
# rstr==3.2.2
# s3transfer==0.10.2
# safetensors==0.4.5
# scikit-learn==1.5.2
# scipy==1.14.1
# shellingham==1.5.4
# simple-websocket==1.0.0
# simpleeval==0.9.13
# six==1.16.0
# sniffio==1.3.1
# snowballstemmer==2.2.0
# SQLAlchemy==2.0.35
# starlette==0.38.5
# sympy==1.13.3
# tenacity==8.5.0
# threadpoolctl==3.5.0
# tiktoken==0.7.0
# tokenizers==0.19.1
# torch==2.4.1
# tqdm==4.66.5
# transformers==4.44.2
# typer==0.12.5
# types-python-dateutil==2.9.0.20240906
# typing-inspect==0.9.0
# typing_extensions==4.12.2
# tzdata==2024.1
# uri-template==1.3.0
# urllib3==2.0.7
# uvicorn==0.30.6
# watchdog==5.0.2
# wcwidth==0.2.13
# webcolors==24.8.0
# Werkzeug==3.0.4
# wrapt==1.16.0
# wsproto==1.2.0
# yarl==1.11.1
# zipp==3.20.2
